// temper/shared-addon-libraries-lib-gps/src/types.ts
export interface WorldSizeMethods {
  Initialize: (this: WorldSizeInstance) => undefined
  GetId: (this: WorldSizeInstance) => number
  GetMapId: (this: WorldSizeInstance) => number
  SetMapId: (this: WorldSizeInstance, id: number) => undefined
  SetZoneId: (this: WorldSizeInstance, zoneId: number) => undefined
  GetZoneId: (this: WorldSizeInstance) => number
  SetSize: (this: WorldSizeInstance, width: number, height: number) => undefined
  GetSize: (this: WorldSizeInstance) => LuaMultiReturn<[number, number]>
  IsValid: (this: WorldSizeInstance) => boolean
  GetName: (this: WorldSizeInstance) => string
  Serialize: (this: WorldSizeInstance) => string
  Deserialize: (this: WorldSizeInstance, data: string) => undefined
}

export interface WorldSizeInstance extends WorldSizeMethods {
  id: number
  zoneId: number
  width: number
  height: number
}

export interface WorldSizeClass extends WorldSizeMethods {
  New: (this: WorldSizeClass) => WorldSizeInstance
}